test(cliniko): revert #58 Swift Testing migration; cross-suite race
CI on the prior commit failed with two races on the process-wide
`URLProtocolStub.currentResponder` singleton:

- `ClinikoStatusThreadingTests.sendWithStatus_returns201_fromCreateTreatmentNote`
  → URLError -2000 (responder cleared mid-flight)
- `ModelDownloaderTests.happyPathDownload`
  → 401 from the real Hugging Face host (responder clobbered, request
  fell through to the network)

`ModelDownloaderTests` already documents this race in its file header:
`.serialized` is **suite-local** in Swift Testing — it scopes the
`URLProtocolStub` singleton cleanly within one suite, but two Swift
Testing suites that both stub HTTP race against each other across the
suite boundary because the parent test scheduler still parallelises
between suites. Adding a second Swift Testing suite (the one this PR
introduced) is exactly that scenario.

Pragmatic resolution for #58:

- Drop `Tests/SpeechToTextTests/Services/Cliniko/ClinikoStatusThreadingTests.swift`.
- Restore the four #58 tests inline alongside their XCTest siblings in
  `ClinikoClientTests.swift` and `TreatmentNoteExporterTests.swift`,
  with a header comment pointing at the race rationale and at the
  follow-up issue. XCTest's scheduling model has been coexisting with
  the Swift Testing `ModelDownloaderTests` without flakes since #20,
  so this restores a known-good arrangement.

The proper structural fix — a process-wide `URLProtocolStubGate` actor
that every HTTP-stubbed test acquires before `install(_:)` — is filed
as a follow-up rather than expanding the scope of this PR.

`swift test --parallel` clean locally (338/338, no failures); CI repro
to follow.
